Separate view and edit project into different components
