+++ client/router.js
@@ -8,7 +8,7 @@ import Home from '@/components/Home';
 import Login from '@/components/auth/Login';
 import NewProject from '@/components/projects/NewProject';
 import Overview from '@/components/Overview';
-import ProjectDetails from '@/components/projects/ProjectDetails';
+import ProjectDetails from '@/components/projects/project-details';
 import Projects from '@/components/projects';
 import Register from '@/components/Register';
 import ResetPassword from '@/components/auth/ResetPassword';
